backend/requirements.txt
Django>=4.0,<5.0
djangorestframework
gunicorn
django-cors-headers
whitenoise
psycopg2-binary
python-dotenv
Pillow
openpyxl
reportlab>=4.0.0
PyPDF2
fpdf
qrcode
django-filter
requests
drf-yasg
python-decouple
Django>=4.0,<5.0
djangorestframework
django-cors-headers
django-filter
drf-yasg
gunicorn
python-dotenv
python-decouple
psycopg2-binary
whitenoise
Pillow
openpyxl
PyPDF2
fpdf
qrcode
requests
djangorestframework-simplejwt==5.3.1


